requirements.txt

anyio==3.6.2
black==22.12.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
fastapi==0.89.1
fonttools==4.38.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
mkl-fft==1.3.1
mkl-service==2.4.0
munkres==1.1.4
mypy-extensions==0.4.3
numpy==1.22.3
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
Pillow==9.4.0
platformdirs==2.6.2
ply==3.11
psycopg2-binary==2.9.5
pydantic==1.10.4
PyQt5==5.15.7
python-multipart==0.0.6
pytz==2022.7.1
sniffio==1.3.0
SQLAlchemy==2.0.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
wincertstore==0.2
